refactor: useProductSearch 훅 이름 변경, validateForm 수정
