+++ src/basic/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from './useDebounce';
 export * from './useNotification';
 export * from './useLocalStorage';
-export * from './useProduct';
+export * from './useProductSearch';
 export * from './useCart';
 export * from './useCoupon';
 export * from './usePayment';
