+++ src/basic/hooks/useProduct.ts
@@ -1,23 +1,49 @@
+import { useCallback, Dispatch, SetStateAction } from 'react';
 import { ProductWithUI } from '../types/product';
 
 interface UseProductPropsType {
-  products: ProductWithUI[];
-  debouncedSearchTerm: string;
+  setProducts: Dispatch<SetStateAction<ProductWithUI[]>>;
+  addNotification: (message: string, type: 'success' | 'error') => void;
 }
 
-export const useProduct = ({ products, debouncedSearchTerm }: UseProductPropsType) => {
-  // 검색어에 따른 상품 필터링
-  const filteredProducts = debouncedSearchTerm
-    ? products.filter(
-        (product) =>
-          product.name.toLowerCase().includes(debouncedSearchTerm.toLowerCase()) ||
-          (product.description &&
-            product.description.toLowerCase().includes(debouncedSearchTerm.toLowerCase()))
-      )
-    : products;
+export const useProduct = ({ setProducts, addNotification }: UseProductPropsType) => {
+  // 상품 추가 (관리자 기능)
+  const addProduct = useCallback(
+    (newProduct: Omit<ProductWithUI, 'id'>) => {
+      const product: ProductWithUI = {
+        ...newProduct,
+        id: `p${Date.now()}`,
+      };
+      setProducts((prev) => [...prev, product]);
+      addNotification(`${newProduct.name} 상품이 추가되었습니다.`, 'success');
+      return product;
+    },
+    [setProducts, addNotification]
+  );
+
+  // 상품 수정 (관리자 기능)
+  const updateProduct = useCallback(
+    (productId: string, updates: Partial<ProductWithUI>) => {
+      setProducts((prev) =>
+        prev.map((product) => (product.id === productId ? { ...product, ...updates } : product))
+      );
+      addNotification(`${updates.name} 상품이 수정되었습니다.`, 'success');
+    },
+    [setProducts, addNotification]
+  );
+
+  // 상품 삭제 (관리자 기능)
+  const deleteProduct = useCallback(
+    (productId: string) => {
+      setProducts((prev) => prev.filter((p) => p.id !== productId));
+      addNotification(`${productId} 상품이 삭제되었습니다.`, 'success');
+    },
+    [setProducts, addNotification]
+  );
 
   return {
-    products,
-    filteredProducts,
+    addProduct,
+    updateProduct,
+    deleteProduct,
   };
 };
