+++ src/basic/utils/validate.ts
@@ -1,6 +1,6 @@
-export const validateForm = (value: string) => {
-  if (value === '' || /^\d+$/.test(value)) {
-    return value === '' ? 0 : parseInt(value);
+export const validateForm = (newValue: string, currentValue: number) => {
+  if (newValue === '' || /^\d+$/.test(newValue)) {
+    return newValue === '' ? 0 : parseInt(newValue);
   }
-  return 0;
+  return currentValue; // 유효하지 않은 입력은 이전 값 유지
 };
